add class builder

+++ xoom/class/File.php
@@ -2,9 +2,12 @@
 
 class File
 {
-    static function create ($filename, $content)
+    static function create ($filename, $content, $overwrite=true)
     {
         $to = Xoom::$rootdir . "/$filename";
+        
+        if (!$overwrite && is_file($to)) return;
+
         file_put_contents($to, $content);
     }
 
@@ -12,5 +15,24 @@ class File
     {
         $files = glob(Xoom::$rootdir . "/$search");
         return $files;
+    }
+
+    static function content ($path)
+    {
+        $content = file_get_contents(Xoom::$rootdir . "/$path");
+        return $content;
+    }
+
+    static function buildClass ($name)
+    {
+        $code = File::content("xoom/etc/ex-Class.php");
+        $dicoas = [
+            "@AUTHOR"   => Config::$adminName       ?? "X",
+            "@LICENCE"  => Config::$adminLicence    ?? "MIT",
+            "@DATETIME" => date("Y-m-d H:i:s"),
+        ];
+
+        $classCode = str_replace(array_keys($dicoas), array_values($dicoas), $code);
+        File::create("xoom/class/$name.php", $classCode, false);
     }
 }
